moved By*ViewBinding helper classes to helper package within the demo app package
