+++ app/build.gradle
@@ -88,6 +88,9 @@ dependencies {
     // slf4j implementation
     implementation 'org.slf4j:slf4j-nop:1.7.25'
 
+    // lifecycle
+    implementation "androidx.lifecycle:lifecycle-common-java8:$lifecycleVersion"
+
     implementation project(path: ':photoeditor')
     implementation project(path: ':stories')
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutinesVersion"
