+++ app/src/main/java/com/automattic/loop/bindinghelpers/ByFragmentViewBinding.kt
@@ -1,4 +1,4 @@
-package com.wordpress.stories
+package com.automattic.loop.bindinghelpers
 
 import android.view.View
 import androidx.fragment.app.Fragment
