+++ app/src/main/java/com/automattic/loop/intro/PermissionRequestFragment.kt
@@ -6,7 +6,7 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import com.automattic.loop.R
 import com.automattic.loop.databinding.FragmentPermissionBinding
-import com.wordpress.stories.viewBinding
+import com.automattic.loop.bindinghelpers.viewBinding
 
 class PermissionRequestFragment : Fragment(R.layout.fragment_permission) {
     interface OnFragmentInteractionListener {
